compat: add module with version-dependant code
Moved there 'base64encode' and 'encode_value' from utils.
Fixed redundant calls in py3 version of 'base64encode'
