+++ torrt/utils.py
@@ -15,6 +15,7 @@ from torrentool.api import Torrent
 
 if False:  # pragma: nocover
     from .base_tracker import GenericTracker
+from .compat import encode_value, base64encode
 
 
 LOGGER = logging.getLogger(__name__)
@@ -61,22 +62,6 @@ def dump_contents(filename, contents):
 
     with open(os.path.join(dump_into, filename), 'wb') as f:
         f.write(text)
-
-
-def base64encode(string_or_bytes):
-    """Return base64 encoded input
-
-    :param string_or_bytes:
-
-    """
-
-    if six.PY3:
-        from base64 import encodebytes
-        string_bytes = string_or_bytes.encode('utf-8') if not isinstance(string_or_bytes, bytes) else string_or_bytes
-        return encodebytes(string_bytes).decode('ascii')
-    else:
-        from base64 import encodestring
-        return encodestring(string_or_bytes)
 
 
 def import_classes():
@@ -264,23 +249,6 @@ def iter_notifiers():
     for notifier_alias, notifier_object in notifier_objects.items():
 
         yield notifier_alias, notifier_object
-
-
-def encode_value(value, encoding=None):
-    """Encodes a value.
-
-    :param str|unicode value:
-    :param str|unicode encoding: Encoding charset.
-    :rtype: bytes
-
-    """
-    if encoding is None:
-        return value
-
-    if six.PY2:
-        value = unicode(value, 'UTF-8')
-
-    return value.encode(encoding)
 
 
 class WithSettings(object):
